Fe research

+++ phd_work/src/make_data/make_bundle_iter_mask.py
@@ -6,22 +6,21 @@ from tqdm import tqdm
 take_log_wf = True
 reco_key = 'reco_ivanov'
 iter_step = 100000
-h5_in = 'pr_q4_14yr_e1_0110_excl_sat_F_excl_geo_F.h5'
-MC_dir_path = '/home3/ivkhar/TA/data/MC/'
-h5_out = h5_in[:-3]+'_bundled.h5'
-h5_in = MC_dir_path + 'merged/composition_spectrum/' + h5_in
-h5_in = '/home/rfit/Telescope_Array/phd_work/data/merged/pr_q4_14yr_e1_0110_excl_sat_F_excl_geo_F.h5'
-MC_dir_path_ROBERT = '/home/rfit/Telescope_Array/phd_work/data/'
-
-os.makedirs('/home/rfit/Telescope_Array/phd_work/data/bundled', exist_ok=True)
+h5_in = 'pr_fe_q4_e1_0110_excl_sat_F_excl_geo_F.h5'
+h5_out = h5_in#[:-3]+'_bundled.h5'
+MC_dir_path_ROBERT = '/home3/rfit/Telescope_Array/phd_work/data/'
+h5_in = os.path.join(MC_dir_path_ROBERT, 'merged/', h5_in)
+os.makedirs('/home3/rfit/Telescope_Array/phd_work/data/bundled', exist_ok=True)
 h5_out = os.path.join(MC_dir_path_ROBERT, 'bundled', h5_out)
 
-take_bdt_params = (0,1,2,5,8,10,11,12,13)
+take_bdt_params = ()#0,1,2,5,8,10,11,12,13)
 take_reco = (0,1,4,5)
 aux_vals = np.array( [-5., -5., 0., 0., 0., 0.] ).astype(np.float32)
 
 proc = 0
-os.remove(h5_out)
+if os.path.exists(h5_out):
+    print('Output file already exists, remove it to continue.')
+    os.remove(h5_out)
 print(os.path.exists(h5_in), h5_out)
 with h5.File(h5_in,'r') as hi, h5.File(h5_out,'w') as ho:
     ev_starts = hi['ev_starts'][:]
@@ -34,11 +33,11 @@ with h5.File(h5_in,'r') as hi, h5.File(h5_out,'w') as ho:
     ho_ds['wfs_flat'] = ho.create_dataset('wfs_flat', shape=np.concatenate( ([num_hits,],hi['dt_wfs'].shape[1:]) ), dtype=hi['dt_wfs'].dtype )
     ho_ds['dt_params'] = ho.create_dataset('dt_params', shape=np.concatenate( ([num_hits,],hi['dt_params'].shape[1:]) ), dtype=hi['dt_params'].dtype )
     ho_ds['recos'] = ho.create_dataset('recos', shape=(num_evs,len(take_reco)+2+len(take_bdt_params)), dtype=hi[reco_key].dtype )
-    ho_ds['dt_bundle'] = ho.create_dataset('dt_bundle', shape=(num_evs,6,6,7), dtype=np.float32 )
+    # ho_ds['dt_bundle'] = ho.create_dataset('dt_bundle', shape=(num_evs,6,6,7), dtype=np.float32 )
     ho_ds['det_max_wf'] = ho.create_dataset('det_max_wf', shape=(num_evs,128,2), dtype=np.float32 )
     ho_ds['det_max_params'] = ho.create_dataset('det_max_params', shape=(num_evs,6), dtype=np.float32 )
     ho_ds['dt_mask'] = ho.create_dataset('dt_mask', shape=(num_hits,2), dtype=np.float32 ) # mask for saturated and geo-excluded dets
-    ho_ds['dt_bunlde_mask'] = ho.create_dataset('dt_bunlde_mask', shape=(num_evs,6,6,2), dtype=np.float32 )
+    # ho_ds['dt_bunlde_mask'] = ho.create_dataset('dt_bunlde_mask', shape=(num_evs,6,6,2), dtype=np.float32 )
     while proc<num_evs:
         print(num_evs, proc)
         step = min(iter_step,num_evs-proc)
@@ -61,7 +60,7 @@ with h5.File(h5_in,'r') as hi, h5.File(h5_out,'w') as ho:
         ho_ds['wfs_flat'][f_hit:l_hit] = wfs
         ho_ds['dt_params'][f_hit:l_hit] = dts
         # сшивка всех параметров иванова + всех параметров рубцова + каких-то bdt_params
-        ho_ds['recos'][f_ev:l_ev] = np.concatenate( (hi[reco_key][f_ev:l_ev][:,take_reco],hi['reco_rubtsov_params'][f_ev:l_ev],hi['bdt_params'][f_ev:l_ev][:,take_bdt_params]), axis=1 )
+        ho_ds['recos'][f_ev:l_ev] = np.concatenate( (hi[reco_key][f_ev:l_ev][:,take_reco],hi['reco_rubtsov_params'][f_ev:l_ev]), axis=1 ) # hi['bdt_params'][f_ev:l_ev][:,take_bdt_params]
         # most active
         idxs_mostQ = np.array([l_ev_starts[i]+np.argmax(dts[l_ev_starts[i]:l_ev_starts[i+1],3]) for i in range(step)])
         mostQ_params = np.array([ dts[idx] for idx in idxs_mostQ ])
@@ -95,10 +94,10 @@ with h5.File(h5_in,'r') as hi, h5.File(h5_out,'w') as ho:
         dt_grid = np.full( (step,6,6,6), aux_vals )
         dt_grid[mask_grid] = dts[mask_flat]
         dt_grid = np.concatenate( (dt_grid,np.expand_dims(mask_grid, axis=-1)), axis=-1 )
-        ho_ds['dt_bundle'][f_ev:l_ev]= dt_grid
+        # ho_ds['dt_bundle'][f_ev:l_ev]= dt_grid
         dt_mask_dets = np.full( (step,6,6,2), False )
         dt_mask_dets[mask_grid] = mask_dets[mask_flat]
-        ho_ds['dt_bunlde_mask'][f_ev:l_ev] = dt_mask_dets
+        # ho_ds['dt_bunlde_mask'][f_ev:l_ev] = dt_mask_dets
         ho_ds['dt_mask'][f_hit:l_hit] = mask_dets
         proc += step
     ho.create_dataset('ev_starts', data=ev_starts)
